Moved FrameSet, FrameSetCollection and _DicomHelper classes into highdicom.utils
